Address some clippy lints

+++ crates/examples/sel4cp/banscii/pds/assistant/core/src/lib.rs
@@ -48,7 +48,7 @@ impl Draft {
         let width = glyphs
             .iter()
             .rev()
-            .map(|g| g.position().x as f32 + g.unpositioned().h_metrics().advance_width)
+            .map(|g| g.position().x + g.unpositioned().h_metrics().advance_width)
             .next()
             .unwrap_or(0.0)
             .ceil() as usize;
